Wye_api/pyproject.toml
[project]
name = "wye-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "email-validator>=2.3.0",
    "langchain>=1.2.15",
    "langchain-google-genai>=2.0.0",
    "langchain-groq>=0.2.0",
    "langchain-openai>=0.3.0",
    "langgraph>=1.1.6",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "PyJWT>=2.10.1",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.44.0",
    "pydantic[email]>=2.12.5",
    "python-socketio>=5.16.1",
]
